For #32: extracted city resolving to BookCreateFragment and Presenter
