+++ app/src/main/java/com/bookcrossing/mobile/location/LocationRepository.kt
@@ -1,3 +1,18 @@
+/*
+ *    Copyright  2019 Andrey Mukamolov
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
 package com.bookcrossing.mobile.location
 
 import android.location.Location
@@ -14,33 +29,36 @@ import io.reactivex.schedulers.Schedulers
 import javax.inject.Inject
 
 class LocationRepository @Inject constructor(
-        private val fusedLocationProviderClient: FusedLocationProviderClient,
-        private val resourceProvider: ResourceProvider,
-        private val localeProvider: LocaleProvider
+  private val fusedLocationProviderClient: FusedLocationProviderClient,
+  private val resourceProvider: ResourceProvider,
+  private val localeProvider: LocaleProvider
 ) {
 
-    fun getLastKnownUserLocation(): Single<Location?> {
-        return Single.create { emitter ->
-            fusedLocationProviderClient.lastLocation
-                    .addOnSuccessListener { emitter.onSuccess(it) }
-                    .addOnFailureListener { emitter.onError(it) }
-        }
+  fun getLastKnownUserLocation(): Single<Location?> {
+    return Single.create { emitter ->
+      fusedLocationProviderClient.lastLocation
+        .addOnSuccessListener { emitter.onSuccess(it) }
+        .addOnFailureListener { emitter.onError(it) }
     }
+  }
 
-    fun resolveUserCity(location: Location): Maybe<String?> {
-        val reverseGeocodeRequest = MapboxGeocoding.builder()
-                .accessToken(resourceProvider.getString(R.string.mapbox_access_token))
-                .languages(localeProvider.currentLocale.language)
-                .limit(1)
-                .query(fromLngLat(location.longitude, location.latitude))
-                .geocodingTypes(GeocodingCriteria.TYPE_PLACE)
-                .build()
+  fun resolveUserCity(location: Location): Maybe<String?> {
+    val reverseGeocodeRequest = MapboxGeocoding.builder()
+      .accessToken(resourceProvider.getString(R.string.mapbox_access_token))
+      .languages(localeProvider.currentLocale.language)
+      .limit(1)
+      .query(fromLngLat(location.longitude, location.latitude))
+      .geocodingTypes(GeocodingCriteria.TYPE_PLACE)
+      .build()
 
-        return Single.fromCallable { reverseGeocodeRequest.executeCall() }
-                .map { response -> response.body()!!.features() }
-                .filter { features -> features.isNotEmpty() }
-                .map { features -> features[0] }
-                .map { feature -> feature.text() }
-                .subscribeOn(Schedulers.io())
-    }
+    return Single.fromCallable { reverseGeocodeRequest.executeCall() }
+      .map { response ->
+        response.body()
+          ?.features()
+      }
+      .filter { features -> features.isNotEmpty() }
+      .map { features -> features[0] }
+      .map { feature -> feature.text() }
+      .subscribeOn(Schedulers.io())
+  }
 }
