+++ app/src/main/res/values-ru/strings.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!--
-  ~    Copyright 2019 Andrey Mukamolov
-  ~
+  ~    Copyright  2019 Andrey Mukamolov
   ~    Licensed under the Apache License, Version 2.0 (the "License");
   ~    you may not use this file except in compliance with the License.
   ~    You may obtain a copy of the License at
@@ -83,7 +82,7 @@
     </string>
   <string name="default_city">Неизвестен</string>
   <string name="title_activity_oss_licenses">Лицензии открытого ПО</string>
-  <string name="enter_city_content">Ошибка получения твоего города. Попробуйте ввести его вручную в настройках, чтобы мы могли автоматически распознавать позицию оставленной книги.</string>
+  <string name="error_enter_city_content">Ошибка получения твоего города. Попробуйте ввести его вручную в настройках, чтобы мы могли автоматически распознавать позицию оставленной книги.</string>
   <string name="enter_city_title">Введите свой город</string>
   <string name="city_hint">Город</string>
   <string name="consent_preference_not_set">Не установлен</string>
